Create .Exercise1
This exercise generates a record vector. Each record saves office information: an identification code, the proprietor's identifier, and the value of expenses. Then it prints the vector elements as they were entered. Subsequently, the program invokes a procedure that sorts the vector based on the office id code. Finally, the generated sorted vector is printed.