Create .Exercise1.pas
This exercise creates a record vector. Each record saves information about offices: the identification code of the office, the proprietor identifier, and the value of expenses. Then, it prints the array information as this was entered. Subsequently, the program invokes a procedure that sorts the original vector. Finally, the generated sorted vector is printed.