Create Exercise2.pas
This exercise works with basic data structures such as arrays and lists. Firstly, it loads an eight-length vector with strings that specify a movie genre in each position. After that, it initializes a list vector with null pointers and invokes a module that loads information about movies: a movie identifier, the code of the genre (an integer number that corresponds with the correlative name in the first string array), and the average score given by the critics. Then it prints the array with the information loaded. Subsequently, a procedure selects the best-ranked movie of each genre and saves it in a new eight-length array (this one with a movie record on each position). Finally, this ultimate vector is sorted and another procedure displays only the first and last position of it (the best and worst-ranked movies).