Create Exercise3.pas
Similar to exercise two, this one too works with basic data structures such as arrays, lists, and the interaction between them. Firstly, it loads a six-length vector with strings that specify a library business branch in each position. After that, it initializes a list vector with null pointers and invokes a module that loads information about products. Then it prints the array with the information loaded. Subsequently, a procedure copies the elements of the third branch in a new array. Finally, this last vector is sorted and another procedure displays it on the console.