Create Exercise4.pas
This program works with the recursive data structure named binary search tree. It implements one procedure for creating a binary tree with integer numbers and functions for the recursive search of the maximum and minimum value inside the tree and one method for searching a value entered by the keyboard.