Create Exercise6.pas
This program works with an array of integer values. The core procedure of it is the binary search method, which uses recursion in order to find a value entered by the keyboard. To achieve this sort the unordered vector first is mandatory, given that the binary search only works with sorted data structures such as arrays with ordered elements or, more commonly, binary search trees.