Create Casting.java
There will be a casting for a TV show. The casting will last a maximum of 5 days and 8 people will be interviewed on different turns each day.
A) Simulate the process of registering people for the casting. Each person is asked for name, ID, and age and must be assigned in one day and turn as follows:
People complete the first day in successive turns, then complete the second day, and so on. Registration ends when a person with the name “ZZZ” arrives, or when the 40 casting slots are filled.
Once registration is complete:
b) Inform the name of the person to be interviewed on each day and turn.